+++ components/Layout/index.tsx
@@ -1,0 +1,3 @@
+export * from './Header';
+export * from './Layout';
+export * from './Nav';
