+++ components/common/index.tsx
@@ -1,7 +1,4 @@
 export * from './Button';
-export * from './Header';
-export * from './Layout';
-export * from './Nav';
 export * from './SearchBar';
 export * from './PartyLogoReplacement';
 export * from './Loading';
