+++ components/index.tsx
@@ -1,3 +1,4 @@
 export * from './Bill';
 export * from './Feed';
 export * from './common';
+export * from './Layout';
